Separated classes; updated references
