+++ StockFighter/OrderDirection.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace StockFighter
+{
+    /// <summary>
+    /// Direction of order.
+    /// </summary>
+    public enum OrderDirection
+    {
+        /// <summary>
+        /// Desire to bid.
+        /// </summary>
+        Buy,
+        /// <summary>
+        /// Desire to ask.
+        /// </summary>
+        Sell
+    }
+}
