+++ StockFighter/Responses/APIResponse.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace StockFighter.Responses
+{
+    /// <summary>
+    /// Base class for all responses. Should be used for commonalities.
+    /// </summary>
+    public class APIResponse
+    {
+        /// <summary>
+        /// Command was received correctly.
+        /// </summary>
+        public bool ok { get; set; }
+        /// <summary>
+        /// Error text. If no error, will be an empty string.
+        /// </summary>
+        public string error { get; set; }
+    }
+}
