+++ StockFighter/Responses/Heartbeat.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace StockFighter.Responses
+{
+    /// <summary>
+    /// Deserialized response from a Heartbeat command.
+    /// </summary>
+    public class Heartbeat : APIResponse { }
+}
